% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwaRs.R
\docType{package}
\name{gwaRs}
\alias{gwaRs}
\alias{gwaRs-package}
\title{Create Manhattan, Q-Q, and PCA plots for GWAS data.}
\description{
A package for creating Manhattan, Q-Q, and PCA plots for GWAS data.
}
\author{
Lindokuhle Nkambule \email{lindonkambule@gmail.com}
}
