% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyotype_plot.R
\name{karyotype_plot}
\alias{karyotype_plot}
\title{Karyotype Plot}
\usage{
karyotype_plot(
  data,
  density.col = c("darkgreen", "yellow", "red"),
  window.size = 1e+06,
  title = NULL
)
}
\arguments{
\item{data}{A data.frame with "CHR" and "BP"columns.}

\item{density.col}{A character vector with colors to use for gradients.}

\item{window.size}{A double precision numeric value indicating the window size.}

\item{title}{A string denoting the title to use for the plot. Default is 'Manhattan Plot'}
}
\value{
A SNP Karyotype plot.
}
\description{
Creates a Karyotype plot
}
\details{
Creates a SNP Karyotype or Density plot from an R dataframe with "CHR" and "BP" columns.
}
\examples{
karyotype_plot(gwasData)

}
\author{
Lindokuhle Nkambule
}
